+++ src/game/stage/inventory-stage.hpp
@@ -159,7 +159,7 @@ namespace stage
         virtual void Draw(sf::RenderTarget & target, const sf::RenderStates &);
         virtual bool KeyRelease(const sf::Event::KeyEvent &);
         virtual void UpdateTime(const float ELAPSED_TIME_SECONDS);
-        virtual void UpdateMousePos(const sf::Vector2f & MOUSE_POS_V);
+        virtual void UpdateMousePos(const sf::Vector2i & MOUSE_POS_V);
         virtual void UpdateMouseDown(const sf::Vector2f & MOUSE_POS_V);
 
     protected:
